# Conflicts:
#	.vs/DuckyData1.0.0Alpha/v14/.suo
#	DuckyData1.0.0Alpha/Content/BugReport/listBugReport.css
#	DuckyData1.0.0Alpha/Content/Site.css
#	DuckyData1.0.0Alpha/Content/VideoFetch/VideoFetchIndex.css
#	DuckyData1.0.0Alpha/Controllers/AccountController.cs
#	DuckyData1.0.0Alpha/Controllers/Manager.cs
#	DuckyData1.0.0Alpha/Controllers/VideoFetchController.cs
#	DuckyData1.0.0Alpha/DuckyData1.0.0Alpha.csproj
#	DuckyData1.0.0Alpha/Factory/Account/AccountFactory.cs
#	DuckyData1.0.0Alpha/Properties/PublishProfiles/DuckyData.pubxml.user
#	DuckyData1.0.0Alpha/Scripts/_references.js
#	DuckyData1.0.0Alpha/Scripts/duckyDataApp.js
#	DuckyData1.0.0Alpha/Scripts/musicFetchCreate.js
#	DuckyData1.0.0Alpha/Scripts/videoFetch.js
#	DuckyData1.0.0Alpha/ViewModels/Account/AccountViewModel.cs
#	DuckyData1.0.0Alpha/Views/Account/Details.cshtml
#	DuckyData1.0.0Alpha/Views/Account/EmailSent.cshtml
#	DuckyData1.0.0Alpha/Views/Account/ListUsers.cshtml
#	DuckyData1.0.0Alpha/Views/BugReports/Create.cshtml
#	DuckyData1.0.0Alpha/Views/BugReports/Edit.cshtml
#	DuckyData1.0.0Alpha/Views/BugReports/Index.cshtml
#	DuckyData1.0.0Alpha/Views/Shared/_Layout.cshtml
#	DuckyData1.0.0Alpha/Views/VideoFetch/Index.cshtml
#	DuckyData1.0.0Alpha/Web.config
#	DuckyData1.0.0Alpha/recognizer.cs

+++ DuckyData1.0.0Alpha/Controllers/MessageController.cs
@@ -1,15 +1,28 @@
 ﻿using AutoMapper;
+using DuckyData1._0._0Alpha;
+using DuckyData1._0._0Alpha.Models;
 using DuckyData1._0._0Alpha.ViewModels;
+using Microsoft.AspNet.Identity;
+using Microsoft.AspNet.Identity.EntityFramework;
 using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+using ACRCloud;
+using static System.Net.WebRequestMethods;
+using System.IO;
 using System.Web.Mvc;
+using DuckyData1._0._0Alpha.Factory.Account;
+using System.Data.Entity.Validation;
 
 namespace DuckyData1._0._0Alpha.Controllers
 {
     public class MessageController : Controller
     {
         private Manager m = new Manager();
+        private AccountFactory ac = new AccountFactory();
+        static private string uID = System.Web.HttpContext.Current.User.Identity.GetUserId();
 
-        
         // GET: Message
         [Authorize(Roles = "Admin")]
         public ActionResult Index()
@@ -71,7 +84,6 @@ namespace DuckyData1._0._0Alpha.Controllers
         {
             if (ModelState.IsValid)
             {
-
                 var addedItem = m.SendMessage(newItem);
                 return RedirectToAction("Inbox", "Message");
             }
@@ -93,9 +105,14 @@ namespace DuckyData1._0._0Alpha.Controllers
         [HttpPost]
         public ActionResult Create(MessageAdd newItem)
         {
+            var user = ac.findUserById(uID);
+            if (user.gagged)
+            {
+                ModelState.AddModelError("SentDate", "Please select an option");
+                return View();
+            }
             if (ModelState.IsValid)
             {
-
                 var addedItem = m.SendMessage(newItem);
                 return RedirectToAction("Inbox", "Message");
             }
@@ -114,10 +131,8 @@ namespace DuckyData1._0._0Alpha.Controllers
             {
                 return HttpNotFound();
             }
-            if (DateTime.Compare(fetchedObject.SentDate, DateTime.Now) > 300)
-            {
-                return RedirectToAction("Inbox", "Message");
-            }
+            //if (DateTime.Compare(fetchedObject.SentDate, DateTime.Now) > 300)
+            
             return View(Mapper.Map<MessageEditForm>(fetchedObject));
             
         }
@@ -184,7 +199,10 @@ namespace DuckyData1._0._0Alpha.Controllers
         {
             if (!id.HasValue) { return HttpNotFound(); }
 
-
+            if( uID != m.GetMessageById(id.Value).UserId && !User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Inbox", "Message");
+            }
             // Attempt to delete the item
             m.DeleteMessage(id.Value);
